requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyyaml
httpx
tenacity
pytest
pytest-asyncio
